timestamp must be a long

+++ src/main/output_modules/raw.py
@@ -21,4 +21,4 @@ def format_event(timestamp, event_type, statuses, settings):
 
 
 def get_timestamp():
-    return time.time() * 1000
+    return int(time.time() * 1000)
